cache-lab: part 1
The part 1 is about writing a simple cache simulator. A cache has 2^s
sets, and each set has E lines. We first extract the set index from the
address, then we compare the tag with every lines in the set. If there's
a matching one, then we got a cache hit, or miss.

If we have a miss, and the set is full, we need evicte a line in that
set. Here we're using LRU as our placement policy, thus we choose doubly
linked list to represent the lines in the set.

Signed-off-by: Jun Zhang <jun@junz.org>